requirements.txt
Sphinx==4.5.0
cmocean==2.0
colorcet==3.0.0
imageio-ffmpeg==0.4.7
imageio>=2.5.0
ipygany==0.5.0
ipyvtklink==0.2.2
jupyter_sphinx==0.4.0
jupyterlab==3.4.3
lxml==4.9.1
matplotlib==3.5.2
meshio==5.3.4
mypy-extensions==0.4.3
mypy==0.961
numpydoc==1.4.0
osmnx==1.2.1
panel==0.13.1
param==1.12.2  # due to panel bug
pydata-sphinx-theme==0.9.0
pypandoc==1.8.1
pytest-sphinx==0.4.0
pythreejs==2.3.0
scipy==1.8.1
sphinx-autobuild==2021.3.14
sphinx-copybutton==0.5.0
sphinx-gallery==0.10.1
sphinx-notfound-page==0.8
sphinx-panels==0.6.0
sphinxcontrib-websupport==1.2.4
trimesh==3.12.7
typed-ast==1.5.4
typing_extensions==4.3.0
sphinx-tabs
sphinxcontrib.asciinema
pyqt5
pyvistaqt
QtPy
